backend/pyproject.toml
[project]
name = "microsoft-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "attr>=0.3.2",
    "boto3>=1.35.0",
    "croniter>=2.0.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "firebase-admin>=6.7.0",
    "langchain-neo4j>=0.4.0",
    "litellm>=1.65.1",
    "mcp>=1.6.0",
    "mem0ai>=0.1.81",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.0",
    "pydantic-ai-slim[tavily]>=0.0.49",
    "pydantic-ai>=0.0.49",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "pytz>=2024.1",
    "rank-bm25>=0.2.2",
    "requests>=2.31.0",
    "smithery>=0.1.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.0",
    "xlsxwriter>=3.2.5",
    "graphiti-core[falkordb,google-genai]>=0.24.1",
    "ddgs>=9.9.2",
]
